Crate axalloc

ArceOS global memory allocator.

It provides GlobalAllocator, which implements the trait core::alloc::GlobalAlloc. A static global variable of type GlobalAllocator is defined with the #[global_allocator] attribute, to be registered as the standard library’s default allocator.

Structs

GlobalAllocator

The global allocator used by ArceOS.

GlobalPage

A RAII wrapper of contiguous 4K-sized pages.

Functions

global_add_memory

Add the given memory region to the global allocator.

global_allocator

Returns the reference to the global allocator.

global_init

Initializes the global allocator with the given memory region.

Type Aliases

DefaultByteAllocator

The default byte allocator.